// packages/cmdl-polymers/src/polymer-types.ts
import { PolymerTreeVisitor } from "./polymer-visitors";
import { CMDLRef } from "cmdl-types";

export interface CMDLPolymerConnection {
  sources: CMDLRef[];
  targets: CMDLRef[];
  quantity: string;
}

export interface CMDLPolymerContainer {
  name: string;
  nodes: CMDLRef[];
  type: string;
  connections?: CMDLPolymerConnection[];
  containers?: CMDLPolymerContainer[];
  parent?: string;
}

export interface CMDLPolymerTree {
  name: string;
  nodes: CMDLRef[];
  connections?: CMDLPolymerConnection[];
  containers?: CMDLPolymerContainer[];
  parent?: string;
}

export interface Serializable {
  /**
   * Converts to string for logging purposes
   */
  print(): string;

  /**
   * Converts to object for export
   */
  toJSON(): any;
}

/**
 * Interface for defining components within the polymer tree
 */
export interface PolymerComponent extends Serializable {
  /**
   * Properties of the particular tree node eg. DP, mol fraction, etc.
   * @TODO improve typing for properties
   */
  properties: Map<string, any>;

  /**
   * Method to determine whether component is a container or entity
   */
  isContainer(): boolean;

  /**
   * Sets provided polymer component as the parent of component calling the method
   * @param arg PolymerComponent
   */
  setParent(arg: PolymerComponent): void;

  /**
   * Generates absolute path for the entity compenent and sets as name
   */
  setName(): void;

  /**
   * Parent component of the node
   */
  parent: PolymerComponent | null;

  /**
   * Name of the node
   */
  name: string;

  /**
   * Accepts a tree visitor for computation of weights
   * @param visitor TreeVisitor
   */
  accept(visitor: PolymerTreeVisitor): void;

  /**
   * Method to serialize the polymer tree into a BigSMILES line notation
   */
  exportToBigSMILES(): string;
}